Fixed duplications between FRY and MEB with the shared directory.
