+++ src/applications/my-education-benefits-apps/fry-dea/validation.js
@@ -1,32 +1,32 @@
-import { isValidEmail } from 'platform/forms/validations';
 import { compareAsc } from 'date-fns';
 import { formFields } from './constants';
+import { isValidPhone, validatePhone } from '../shared/validations';
 
-const isValidPhone = (phone, isInternational) => {
-  let stripped;
-  try {
-    stripped = phone.replace(/[^\d]/g, '');
-  } catch (err) {
-    stripped = phone;
-  }
-  return isInternational
-    ? /^\d{10,15}$/.test(stripped)
-    : /^\d{10}$/.test(stripped);
-};
+// const isValidPhone = (phone, isInternational) => {
+//   let stripped;
+//   try {
+//     stripped = phone.replace(/[^\d]/g, '');
+//   } catch (err) {
+//     stripped = phone;
+//   }
+//   return isInternational
+//     ? /^\d{10,15}$/.test(stripped)
+//     : /^\d{10}$/.test(stripped);
+// };
 
 export const isValidPhoneField = phoneField => {
   const { isInternational } = phoneField;
   return isValidPhone(phoneField.phone, isInternational);
 };
 
-const validatePhone = (errors, phone, isInternational) => {
-  if (phone && !isValidPhone(phone, isInternational)) {
-    const numDigits = isInternational ? '10 to 15' : '10';
-    errors.addError(
-      `Please enter a ${numDigits}-digit phone number (with or without dashes)`,
-    );
-  }
-};
+// const validatePhone = (errors, phone, isInternational) => {
+//   if (phone && !isValidPhone(phone, isInternational)) {
+//     const numDigits = isInternational ? '10 to 15' : '10';
+//     errors.addError(
+//       `Please enter a ${numDigits}-digit phone number (with or without dashes)`,
+//     );
+//   }
+// };
 
 export const validateHomePhone = (errors, phone, formData) => {
   const { isInternational } = formData[formFields.viewPhoneNumbers][
@@ -43,11 +43,11 @@ export const validateMobilePhone = (errors, phone, formData) => {
   validatePhone(errors, phone, isInternational);
 };
 
-export const validateEmail = (errors, email) => {
-  if (email && !isValidEmail(email)) {
-    errors.addError('Please enter a valid email address.');
-  }
-};
+// export const validateEmail = (errors, email) => {
+//   if (email && !isValidEmail(email)) {
+//     errors.addError('Please enter a valid email address.');
+//   }
+// };
 
 export const validateReMarriageDate = (errors, newMarriageDate, formData) => {
   const result = compareAsc(
