+++ src/applications/my-education-benefits-apps/my-education-benefits/utils/validation.js
@@ -1,27 +1,28 @@
-import { isValidEmail } from 'platform/forms/validations';
+// import { isValidEmail } from 'platform/forms/validations';
 import moment from 'moment';
 import { formatReadableDate } from '../helpers';
+import { validatePhone } from '../../shared/validations';
 
-export const isValidPhone = (phone, isInternational) => {
-  let stripped;
-  try {
-    stripped = phone.replace(/[^\d]/g, '');
-  } catch (err) {
-    stripped = phone;
-  }
-  return isInternational
-    ? /^\d{10,15}$/.test(stripped)
-    : /^\d{10}$/.test(stripped);
-};
+// export const isValidPhone = (phone, isInternational) => {
+//   let stripped;
+//   try {
+//     stripped = phone.replace(/[^\d]/g, '');
+//   } catch (err) {
+//     stripped = phone;
+//   }
+//   return isInternational
+//     ? /^\d{10,15}$/.test(stripped)
+//     : /^\d{10}$/.test(stripped);
+// };
 
-const validatePhone = (errors, phone, isInternational) => {
-  if (phone && !isValidPhone(phone, isInternational)) {
-    const numDigits = isInternational ? '10 to 15' : '10';
-    errors.addError(
-      `Please enter a ${numDigits}-digit phone number (with or without dashes)`,
-    );
-  }
-};
+// const validatePhone = (errors, phone, isInternational) => {
+//   if (phone && !isValidPhone(phone, isInternational)) {
+//     const numDigits = isInternational ? '10 to 15' : '10';
+//     errors.addError(
+//       `Please enter a ${numDigits}-digit phone number (with or without dashes)`,
+//     );
+//   }
+// };
 
 export const validateHomePhone = (errors, phone, formData) => {
   const { isInternational } = formData['view:phoneNumbers'].phoneNumber;
@@ -33,11 +34,11 @@ export const validateMobilePhone = (errors, phone, formData) => {
   validatePhone(errors, phone, isInternational);
 };
 
-export const validateEmail = (errors, email) => {
-  if (email && !isValidEmail(email)) {
-    errors.addError('Please enter a valid email address.');
-  }
-};
+// export const validateEmail = (errors, email) => {
+//   if (email && !isValidEmail(email)) {
+//     errors.addError('Please enter a valid email address.');
+//   }
+// };
 
 export const validateEffectiveDate = (errors, dateString) => {
   const effectiveDate = moment(dateString);
