+++ src/applications/my-education-benefits-apps/shared/schemas.js
@@ -1,0 +1,18 @@
+import commonDefinitions from 'vets-json-schema/dist/definitions.json';
+
+const { usaPhone } = commonDefinitions;
+
+export function phoneSchema() {
+  return {
+    type: 'object',
+    properties: {
+      phone: {
+        ...usaPhone,
+        pattern: '^\\d[-]?\\d(?:[0-9-]*\\d)?$',
+      },
+      isInternational: {
+        type: 'boolean',
+      },
+    },
+  };
+}
