+++ src/applications/my-education-benefits-apps/shared/validations.js
@@ -1,0 +1,28 @@
+import { isValidEmail } from 'platform/forms/validations';
+
+export const isValidPhone = (phone, isInternational) => {
+  let stripped;
+  try {
+    stripped = phone.replace(/[^\d]/g, '');
+  } catch (err) {
+    stripped = phone;
+  }
+  return isInternational
+    ? /^\d{10,15}$/.test(stripped)
+    : /^\d{10}$/.test(stripped);
+};
+
+export const validatePhone = (errors, phone, isInternational) => {
+  if (phone && !isValidPhone(phone, isInternational)) {
+    const numDigits = isInternational ? '10 to 15' : '10';
+    errors.addError(
+      `Please enter a ${numDigits}-digit phone number (with or without dashes)`,
+    );
+  }
+};
+
+export const validateEmail = (errors, email) => {
+  if (email && !isValidEmail(email)) {
+    errors.addError('Please enter a valid email address.');
+  }
+};
